receipt/requirements.txt
# Base dependencies
numpy>=1.18.5
opencv-python>=4.1.1
pillow>=9.5.0
requests>=2.32.0
torch>=1.7.0
torchvision>=0.8.1
ultralytics>=8.2.34  # YOLOv5
pytesseract>=0.3.10
Flask>=2.2.0,<2.3.0

# Additional
pandas>=1.1.4
